moving type definitions to separate file
